Adjust final page detection for incrementals
Previously, we were trying to have our last page return 0 rows.

With incremental updates, this meant that we were indexing lots of
short pages, because by the time one query ran, activity in the
database would have resulted in a small number of items being in the
next page, etc.

This removes a previous workaround to stop an incremental run when
the number of records was less than 10.

Signed-off-by: Jeff Godin <jgodin@tadl.org>