Add validation to sort_year
Add validation to the sort_year field -- we only want integers, and
dateIssued with encoding="marc" can still contain values like
'199u'.

In these cases, we omit the value. Normalizing to 1990 may be
another option. Sorting exact years and approximate years may not be
something that we can present in an intuitive fashion.

Signed-off-by: Jeff Godin <jgodin@tadl.org>